Outdated Sequences description fix
I'm starting by reading `datamodel.rst` from the start, to not miss anything. "Sequences" should be more linked into, or its info needs repetition. (Going from links in `__getitem__`, it's unreachable.)

"Some" (sequences) is inaccurate today.

`a[-n]` doesn't exist, and it should be clear. Also, I changed "2nd to last" example in order to make the English simpler.

I'm renaming variables in slicing. *i* was used as an index, but now we are jumping to *k*? And then to *x*? 

The expression "`e<=i<f`" is false for negatives, and "The comment above" is putting reader into the previous paragraph.

"extended slicing" is an old thing. From simple google search, it was outdated 12 years ego. It's just slicing today.

Again negatives and negative step. 

Parameters can be omitted.

The line about distinguishment by mutability is pointless. And/or with with to small font.